Added Recoil, dark theme, and i18n support
Wrapped App component as Recoil is not currently used but will be needed in the future.
Dark theme currently uses useContext and createContext, but will be controlled by Recoil in the future. This is also in the form of wrapping the App component.
i18n is located in the . /src/locales folder to support language changes by reading files defined in the /src/locales folder.