Add Button component to be displayed in header
The display and some processing of the three button components to be displayed in the header are defined in buttons.tsx.
Theme button: This button switches the page theme according to the color theme defined in theme.tsx. The display of the button is also changed according to the theme.
Language conversion button: Opens a modal when pressed, invokes the language conversion instance defined in locales, and is responsible for converting the page text to the selected language.
connectWallet button: At present, nothing happens when this button is pressed because ethers.js, etc. are not used, but it is being prepared as an entry point to connect to chains such as Ethereum in the future.