Add p5 instances to be displayed in the image component
Although commented out, sketch.ts will be the recipient of logic that processes data retrieved via WebAPI and converts unique values to p5, and sketch.ts will be responsible for the process of applying the received values to the p5 instance. Currently, a sample p5 instance is directly written in sketch.ts for the sake of provisional use.